cmd: move function copyFile into utils.go
